backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg[binary]==3.2.1
bleach==6.2.0
gunicorn==22.0.0
Pillow==10.4.0
django-redis==5.4.0
